Code/ai-study-pal-ui/backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pandas==2.2.3
matplotlib==3.7.2
scikit-learn==1.3.2

nltk==3.8.1
numpy==1.26.4
pypdf==3.10.0
python-docx==0.8.11
networkx==3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.4.0
# Optional (needed for CrossEncoder / local HF models):
transformers==4.35.0
torch==2.2.0
python-dotenv==1.0.0
